+++ objc2/src/message/gnustep.rs
@@ -1,14 +1,12 @@
 use core::mem;
 
-use super::{conditional_try, Encode, MessageArguments, MessageError};
+use super::conditional_try;
 use crate::ffi;
 use crate::runtime::{Class, Object, Sel};
+use crate::{Encode, MessageArguments};
 
-pub(crate) unsafe fn send_unverified<A, R>(
-    receiver: *mut Object,
-    sel: Sel,
-    args: A,
-) -> Result<R, MessageError>
+#[track_caller]
+pub(crate) unsafe fn send_unverified<A, R>(receiver: *mut Object, sel: Sel, args: A) -> R
 where
     A: MessageArguments,
     R: Encode,
@@ -28,12 +26,13 @@ where
     unsafe { conditional_try(|| A::__invoke(msg_send_fn, receiver, sel, args)) }
 }
 
+#[track_caller]
 pub(crate) unsafe fn send_super_unverified<A, R>(
     receiver: *mut Object,
     superclass: &Class,
     sel: Sel,
     args: A,
-) -> Result<R, MessageError>
+) -> R
 where
     A: MessageArguments,
     R: Encode,