+++ objc2/CHANGELOG.md
@@ -37,7 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `Class::verify_sel`.
 
 ### Changed
-* **BREAKING:** `Sel` is now required to be non-null, which means that you
+* **BREAKING**: `Sel` is now required to be non-null, which means that you
   have to ensure that any selectors you receive from method calls are
   non-null before using them.
 * **BREAKING**: `ClassBuilder::root` is now generic over the function pointer,
@@ -46,9 +46,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * **BREAKING**: Moved `MessageReceiver::verify_message` to `Class::verify_sel`
   and changed return type.
 * Improved debug output with `verify_message` feature enabled.
+* **BREAKING**: Changed `MessageReceiver::send_message` to panic instead of
+  returning an error.
 
 ### Removed
-* **BREAKING:** Removed the `Sel::from_ptr` and `Sel::as_ptr` methods.
+* **BREAKING**: Removed the `Sel::from_ptr` and `Sel::as_ptr` methods.
+* **BREAKING**: Removed `MessageError`.
 
 
 ## 0.3.0-beta.0 - 2022-06-13